+++ packages/sdks/angular-sdk/projects/angular-sdk/src/environment.ts
@@ -1,0 +1,3 @@
+export const environment = {
+  buildVersion: '0.5.10'
+};