+++ packages/sdks/angular-sdk/projects/demo-app/src/app/home/home.component.spec.ts
@@ -1,0 +1,44 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { HomeComponent } from './home.component';
+import createSdk from '@descope/web-js-sdk';
+import mocked = jest.mocked;
+import { NO_ERRORS_SCHEMA } from '@angular/core';
+import { DescopeAuthConfig } from '../../../../angular-sdk/src/lib/types/types';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+
+jest.mock('@descope/web-js-sdk');
+
+describe('HomeComponent', () => {
+  let component: HomeComponent;
+  let fixture: ComponentFixture<HomeComponent>;
+
+  let mockedCreateSdk: jest.Mock;
+  const onSessionTokenChangeSpy = jest.fn();
+  const onUserChangeSpy = jest.fn();
+
+  beforeEach(() => {
+    mockedCreateSdk = mocked(createSdk);
+    mockedCreateSdk.mockReturnValue({
+      onSessionTokenChange: onSessionTokenChangeSpy,
+      onUserChange: onUserChangeSpy
+    });
+
+    TestBed.configureTestingModule({
+      imports: [HttpClientTestingModule],
+      schemas: [NO_ERRORS_SCHEMA],
+      declarations: [HomeComponent],
+      providers: [
+        DescopeAuthConfig,
+        { provide: DescopeAuthConfig, useValue: { projectId: 'test' } }
+      ]
+    });
+    fixture = TestBed.createComponent(HomeComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});