+++ packages/sdks/angular-sdk/projects/demo-app/src/environments/conifg.ts
@@ -1,0 +1,14 @@
+export interface Env {
+  descopeProjectId: string;
+  descopeBaseUrl?: string;
+  descopeBaseStaticUrl?: string;
+  descopeFlowId?: string;
+  descopeDebugMode?: false;
+  descopeTheme?: string;
+  descopeLocale?: string;
+  descopeRedirectUrl?: string;
+  descopeTenantId?: string;
+  descopeTelemetryKey?: string;
+  descopeStepUpFlowId?: string;
+  backendUrl?: string;
+}